Refactor NZBProvider.py -> nzb_provider.py
