+++ medusa/providers/nzb/anizb.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 import traceback
 
 from requests.compat import urljoin
-from .NZBProvider import NZBProvider
+from .nzb_provider import NZBProvider
 from ... import logger, tv_cache
 from ...bs4_parser import BS4Parser
 from ...helper.common import try_int
