+++ medusa/providers/nzb/binsearch.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 import re
 
 from requests.compat import urljoin
-from .NZBProvider import NZBProvider
+from .nzb_provider import NZBProvider
 from ... import logger, tv_cache
