+++ medusa/providers/nzb/womble.py
@@ -19,7 +19,7 @@
 from __future__ import unicode_literals
 
 from requests.compat import urljoin
-from .NZBProvider import NZBProvider
+from .nzb_provider import NZBProvider
 from ... import logger, tv_cache
