+++ tests/legacy/providers/nzb_provider_tests.py
@@ -22,7 +22,7 @@ import os
 
 import medusa as app
 from medusa.providers.generic_provider import GenericProvider
-from medusa.providers.nzb.NZBProvider import NZBProvider
+from medusa.providers.nzb.nzb_provider import NZBProvider
 from six import iteritems
 from .generic_provider_tests import GenericProviderTests
